Move shared value parsing logic into helper class
