+++ core/src/test/java/com/graphhopper/routing/util/DataFlagEncoderTest.java
@@ -23,12 +23,12 @@ import com.graphhopper.routing.profiles.RoadClass;
 import com.graphhopper.routing.profiles.RoadEnvironment;
 import com.graphhopper.routing.profiles.Surface;
 import com.graphhopper.routing.util.parsers.OSMMaxSpeedParser;
-import com.graphhopper.routing.util.parsers.OSMMaxWidthParser;
 import com.graphhopper.routing.util.parsers.OSMRoadAccessParser;
 import com.graphhopper.routing.util.parsers.OSMRoadClassParser;
 import com.graphhopper.routing.util.parsers.OSMRoadEnvironmentParser;
 import com.graphhopper.routing.util.parsers.OSMSurfaceParser;
 import com.graphhopper.routing.util.parsers.SpatialRuleParser;
+import com.graphhopper.routing.util.parsers.helpers.OSMDimensionExtractor;
 import com.graphhopper.routing.util.spatialrules.SpatialRule;
 import com.graphhopper.routing.util.spatialrules.SpatialRuleLookup;
 import com.graphhopper.routing.util.spatialrules.countries.GermanySpatialRule;
@@ -322,34 +322,34 @@ public class DataFlagEncoderTest {
 
     @Test
     public void stringToMeter() {
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("1.5"), DELTA);
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("1.5m"), DELTA);
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("1.5 m"), DELTA);
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("1.5   m"), DELTA);
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("1.5 meter"), DELTA);
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("4 ft 11 in"), DELTA);
-        assertEquals(1.5, OSMMaxWidthParser.stringToMeter("4'11''"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("1.5"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("1.5m"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("1.5 m"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("1.5   m"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("1.5 meter"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("4 ft 11 in"), DELTA);
+        assertEquals(1.5, OSMDimensionExtractor.stringToMeter("4'11''"), DELTA);
 
 
-        assertEquals(3, OSMMaxWidthParser.stringToMeter("3 m."), DELTA);
-        assertEquals(3, OSMMaxWidthParser.stringToMeter("3meters"), DELTA);
-        assertEquals(0.8 * 3, OSMMaxWidthParser.stringToMeter("~3"), DELTA);
-        assertEquals(3 * 0.8, OSMMaxWidthParser.stringToMeter("3 m approx"), DELTA);
+        assertEquals(3, OSMDimensionExtractor.stringToMeter("3 m."), DELTA);
+        assertEquals(3, OSMDimensionExtractor.stringToMeter("3meters"), DELTA);
+        assertEquals(0.8 * 3, OSMDimensionExtractor.stringToMeter("~3"), DELTA);
+        assertEquals(3 * 0.8, OSMDimensionExtractor.stringToMeter("3 m approx"), DELTA);
 
         // 2.743 + 0.178
-        assertEquals(2.921, OSMMaxWidthParser.stringToMeter("9 ft 7in"), DELTA);
-        assertEquals(2.921, OSMMaxWidthParser.stringToMeter("9'7\""), DELTA);
-        assertEquals(2.921, OSMMaxWidthParser.stringToMeter("9'7''"), DELTA);
-        assertEquals(2.921, OSMMaxWidthParser.stringToMeter("9' 7\""), DELTA);
+        assertEquals(2.921, OSMDimensionExtractor.stringToMeter("9 ft 7in"), DELTA);
+        assertEquals(2.921, OSMDimensionExtractor.stringToMeter("9'7\""), DELTA);
+        assertEquals(2.921, OSMDimensionExtractor.stringToMeter("9'7''"), DELTA);
+        assertEquals(2.921, OSMDimensionExtractor.stringToMeter("9' 7\""), DELTA);
 
-        assertEquals(2.743, OSMMaxWidthParser.stringToMeter("9'"), DELTA);
-        assertEquals(2.743, OSMMaxWidthParser.stringToMeter("9 feet"), DELTA);
+        assertEquals(2.743, OSMDimensionExtractor.stringToMeter("9'"), DELTA);
+        assertEquals(2.743, OSMDimensionExtractor.stringToMeter("9 feet"), DELTA);
     }
 
     @Test(expected = NumberFormatException.class)
     public void stringToMeterException() {
         // Unexpected values
-        OSMMaxWidthParser.stringToMeter("height limit 1.5m");
+        OSMDimensionExtractor.stringToMeter("height limit 1.5m");
     }
 
     @Test
